Adding manual node restart

+++ simplyblock_core/controllers/snapshot_controller.py
@@ -170,8 +170,8 @@ def delete(snapshot_uuid, force=False):
     snap.remove(db_controller.kv_store)
 
     base_lvol = db_controller.get_lvol_by_id(snap.lvol.get_id())
-    if base_lvol.deleted is True:
-        lvol_controller.delete_lvol(base_lvol.get_id())
+    if base_lvol and base_lvol.deleted is True:
+            lvol_controller.delete_lvol(base_lvol.get_id())
 
     logger.info("Done")
     snapshot_events.snapshot_delete(snap)
